requirements/dev.txt
-r base.txt

# Development tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0
Faker==24.4.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==26.3.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.5.0
